+++ src/temple_refactor1/scripts/index.js
@@ -1,4 +1,3 @@
-
 import "../styles/test.scss";
 import { init } from "./bootstrap";
